+++ packages/tendermint-rpc/src/client.spec.ts
@@ -485,6 +485,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, appCr
         expect(event.height).toBeGreaterThan(0);
         expect(event.index).toEqual(0);
         expect(event.result).toBeTruthy();
+        expect(event.result.events.length).toBeGreaterThanOrEqual(1);
 
         events.push(event);
 
@@ -507,12 +508,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, appCr
     expect(events.length).toEqual(2);
     // Meta
     expect(events[1].height).toEqual(events[0].height + 1);
-    if (events[1].result.tags && events[0].result.tags) {
-      expect(events[1].result.tags).not.toEqual(events[0].result.tags);
-    }
-    if (events[1].result.events && events[0].result.events) {
-      expect(events[1].result.events).not.toEqual(events[0].result.events);
-    }
+    expect(events[1].result.events).not.toEqual(events[0].result.events);
     // Content
     expect(events[0].tx).toEqual(transactionData1);
     expect(events[1].tx).toEqual(transactionData2);
@@ -536,6 +532,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, appCr
         expect(event.height).toBeGreaterThan(0);
         expect(event.index).toEqual(0);
         expect(event.result).toBeTruthy();
+        expect(event.result.events.length).toBeGreaterThanOrEqual(1);
         events.push(event);
 
         if (events.length === 2) {
@@ -554,12 +551,7 @@ function websocketTestSuite(rpcFactory: () => RpcClient, adaptor: Adaptor, appCr
     expect(events.length).toEqual(2);
     // Meta
     expect(events[1].height).toEqual(events[0].height + 1);
-    if (events[1].result.tags && events[0].result.tags) {
-      expect(events[1].result.tags).not.toEqual(events[0].result.tags);
-    }
-    if (events[1].result.events && events[0].result.events) {
-      expect(events[1].result.events).not.toEqual(events[0].result.events);
-    }
+    expect(events[1].result.events).not.toEqual(events[0].result.events);
     // Content
     expect(events[0].tx).toEqual(transactionData1);
     expect(events[1].tx).toEqual(transactionData2);