+++ packages/tendermint-rpc/src/responses.ts
@@ -177,8 +177,7 @@ export interface TxData {
   readonly code: number;
   readonly log?: string;
   readonly data?: Uint8Array;
-  readonly tags?: readonly Tag[];
-  readonly events?: readonly Event[];
+  readonly events: readonly Event[];
   // readonly fees?: any;
 }