+++ CHANGELOG.md
@@ -54,6 +54,8 @@
   is always set.
 - @cosmjs/tendermint-rpc: Change type of `GenesisResponse.appState` to
   `Record<string, unknown> | undefined`.
+- @cosmjs/tendermint-rpc: Remove obsolete `TxData.tags` and make `TxData.events`
+  non-optional.
 - @cosmjs/utils: Add `assertDefined`.
 - @cosmjs/faucet: Rename binary from `cosmwasm-faucet` to `cosmos-faucet`.